WIP reorganizing fill regions to be in region_fill and solver setup
